Extract Examples
Previously, the Examples class represented "typical" usage of the client, and
was a good test to indicate that nothing was broken.  This change extracts
those tests into a JUnit 5 Test Interface allowing them to be reused across
multiple database types.
